--- server/kolide/emails_test.go
@@ -1,19 +1,0 @@
-package kolide
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestTemplateProcessor(t *testing.T) {
-	mailer := PasswordResetMailer{
-		BaseURL: "https://localhost.com:8080",
-		Token:   "12345",
-	}
-
-	out, err := mailer.Message()
-	require.Nil(t, err)
-	assert.NotNil(t, out)
-}
