+++ server/kolide/invites.go
@@ -1,9 +1,7 @@
 package kolide
 
 import (
-	"bytes"
 	"context"
-	"html/template"
 )
 
 // InviteStore contains the methods for
@@ -70,26 +68,4 @@ type Invite struct {
 	Position   string `json:"position,omitempty"`
 	Token      string `json:"-"`
 	SSOEnabled bool   `json:"sso_enabled" db:"sso_enabled"`
-}
-
-// InviteMailer is used to build an email template for the invite email.
-type InviteMailer struct {
-	*Invite
-	BaseURL           template.URL
-	AssetURL          template.URL
-	InvitedByUsername string
-	OrgName           string
-}
-
-func (i *InviteMailer) Message() ([]byte, error) {
-	t, err := getTemplate("server/mail/templates/invite_token.html")
-	if err != nil {
-		return nil, err
-	}
-
-	var msg bytes.Buffer
-	if err = t.Execute(&msg, i); err != nil {
-		return nil, err
-	}
-	return msg.Bytes(), nil
 }
