+++ server/service/service_appconfig.go
@@ -64,7 +64,7 @@ func (svc service) SendTestEmail(ctx context.Context, config *kolide.AppConfig) 
 	testMail := kolide.Email{
 		Subject: "Hello from Fleet",
 		To:      []string{vc.User.Email},
-		Mailer: &kolide.SMTPTestMailer{
+		Mailer: &mail.SMTPTestMailer{
 			BaseURL:  template.URL(config.KolideServerURL + svc.config.Server.URLPrefix),
 			AssetURL: getAssetURL(),
 		},
